- 13-11-2004 fixed: Alpha level errors that occured with some button types (radio button etc.) + fixed a compiler warning
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@2408 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/CdgParser.cpp
@@ -100,7 +100,7 @@ void CCdgLoader::OnStartup()
 		m_CdgFileState = FILE_ERR_OPENING;
 		return;
 	}
-	m_uiFileLength = m_File.GetLength();
+	m_uiFileLength = (int)m_File.GetLength();	// ASSUMES FILELENGTH IS LESS THAN 2^32 bytes!!!
 	if(!m_uiFileLength)	return;
 	m_File.Seek(0,SEEK_SET);
 	if(m_pBuffer)
